// src/app/dometa-api/model/entityRelationModel.ts
/**
 * DoMeta API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { MetaTypInfoModel } from './metaTypInfoModel';

export interface EntityRelationModel { 
    name?: string;
    metaType?: MetaTypInfoModel;
    minimum?: number;
    maximum?: number;
}